requirements.txt
python-igraph
connexion[swagger-ui]
python_dateutil
pymongo
hashids
redis
shyaml
celery
gunicorn
